requirements.txt
Django==3.2.23
django-tenants==3.6.1
psycopg2-binary==2.9.5
django-tenant-schemas==1.11.0
django-extensions==3.2.3
bpython==0.24
stripe==7.7.0
cloudinary==1.37.0
pillow==10.1.0
django-elasticsearch-dsl==7.4
elasticsearch==7.17
elasticsearch-dsl==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-richtextfield==1.6.1
django-sendgrid-v5==1.2.3
redis==5.0.1
celery==5.3.6